+++ front/lib/api/actions/servers/project_manager/index.ts
@@ -12,7 +12,7 @@ import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
  * This server provides tools for managing projects:
  * - File operations: list, add, and update project files
  * - Metadata operations: edit description, add/edit URLs
- * - Search operations: search unread conversations
+ * - Retrieval: recent documents from the project data source and context nodes
  */
 function createServer(
   auth: Authenticator,
